Move tools directory within src
