+++ src/tools/bootstrapAdminCLI.js
